main

+++ database/migrations/2024_03_06_063647_create_hotel_facilities_table.php
@@ -13,17 +13,19 @@ return new class extends Migration
      */
     public function up()
     {
-        Schema::create('hotel_facilities', function (Blueprint $table) {
-            $table->id();
-            $table->string('name')->nullable();
-            $table->text('description')->nullable();
-            $table->unsignedBigInteger('hotel_id')->nullable();
-            $table->foreign('hotel_id')->references('id')->on('hotels')->onDelete('cascade');
-            $table->tinyInteger('status')->nullable()->default(1);
-            $table->text('image')->nullable();
+        if (!Schema::hasTable('hotel_facilities')) {
+            Schema::create('hotel_facilities', function (Blueprint $table) {
+                $table->id();
+                $table->string('name')->nullable();
+                $table->text('description')->nullable();
+                $table->unsignedBigInteger('hotel_id')->nullable();
+                $table->foreign('hotel_id')->references('id')->on('hotels')->onDelete('cascade');
+                $table->tinyInteger('status')->nullable()->default(1);
+                $table->text('image')->nullable();
 
-            $table->timestamps();
-        });
+                $table->timestamps();
+            });
+        }
     }
 
     /**
